// apps/admin/src/generated/api/schemas/positionListResponseDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */
import type { PositionResponseDto } from "./positionResponseDto";

export interface PositionListResponseDto {
  items: PositionResponseDto[];
  total: number;
  page: number;
  pageSize: number;
}